refactor(saves): address review on situational saving throw roll modes (#401)
Share one saving throw roll mode calculator between the character creation
dialog and the saving throw config, name Highborn in the rule reference,
scope Migration047's already-present check to the poison rule, demote
formatRollModeLabel back into its only caller, and token the artwork backdrop.
