+++ src/scss/components/_document-card.scss
@@ -227,7 +227,7 @@
     flex-shrink: 0;
     border: 0;
     border-radius: 0;
-    background-color: rgba(0, 0, 0, 0.7);
+    background-color: var(--nimble-artwork-backdrop);
     aspect-ratio: 1 / 1;
     object-fit: cover;
     object-position: center;
